+++ CHANGELOG.md
@@ -4,4 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.0] - 2020-10-16
+### Added
+- Ability to create a custom `fit` command extension via the `FitModel` trait.
+- Ability to create a custom `score` command extension via the `ScoreModel` trait.
+- Ability to create a custom `apply` command extension via the `ApplyModel` trait.
+- This CHANGELOG
+- Makefile
+- License